split actors into multiple files
